requirements.txt
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.2

django-js-asset==2.2.0
et-xmlfile==1.1.0
idna==3.6
openpyxl==3.1.2
pillow==10.2.0
psycopg2-binary==2.9.9
python-decouple==3.8
requests==2.31.0
sqlparse==0.4.4
tzdata==2023.4
urllib3==2.2.0
